rebased, refactored
