+++ package.json
@@ -11,7 +11,6 @@
     "http-proxy-agent": "^4.0.1",
     "https-proxy-agent": "^5.0.0",
     "jose": "^2.0.2",
-    "jsonwebtoken": "^8.5.1",
     "limiter": "^1.1.5",
     "lru-memoizer": "^2.1.2",
     "ms": "^2.1.2",
@@ -32,6 +31,7 @@
     "eslint-plugin-babel": "^5.3.0",
     "express": "^4.17.1",
     "express-jwt": "^6.0.0",
+    "jsonwebtoken": "^8.5.1",
     "koa": "^2.12.1",
     "koa-jwt": "^3.6.0",
     "mocha": "^6.2.3",
