+++ src/utils.js
@@ -1,86 +1,20 @@
-export function certToPEM(cert) {
-  cert = cert.match(/.{1,64}/g).join('\n');
-  cert = `-----BEGIN CERTIFICATE-----\n${cert}\n-----END CERTIFICATE-----\n`;
-  return cert;
-}
-
-function prepadSigned(hexStr) {
-  const msb = hexStr[0];
-  if (msb < '0' || msb > '7') {
-    return `00${hexStr}`;
-  }
-  return hexStr;
-}
-
-function toHex(number) {
-  const nstr = number.toString(16);
-  if (nstr.length % 2) {
-    return `0${nstr}`;
-  }
-  return nstr;
-}
-
-function encodeLengthHex(n) {
-  if (n <= 127) {
-    return toHex(n);
-  }
-  const nHex = toHex(n);
-  const lengthOfLengthByte = 128 + nHex.length / 2;
-  return toHex(lengthOfLengthByte) + nHex;
-}
-
-/*
- * Source: http://stackoverflow.com/questions/18835132/xml-to-pem-in-node-js
- */
-export function rsaPublicKeyToPEM(modulusB64, exponentB64) {
-  const modulus = Buffer.from(modulusB64, 'base64');
-  const exponent = Buffer.from(exponentB64, 'base64');
-  const modulusHex = prepadSigned(modulus.toString('hex'));
-  const exponentHex = prepadSigned(exponent.toString('hex'));
-  const modlen = modulusHex.length / 2;
-  const explen = exponentHex.length / 2;
-
-  const encodedModlen = encodeLengthHex(modlen);
-  const encodedExplen = encodeLengthHex(explen);
-  const encodedPubkey = '30' +
-    encodeLengthHex(modlen + explen + encodedModlen.length / 2 + encodedExplen.length / 2 + 2) +
-    '02' + encodedModlen + modulusHex +
-    '02' + encodedExplen + exponentHex;
-
-  const der = Buffer.from(encodedPubkey, 'hex')
-    .toString('base64');
-
-  let pem = '-----BEGIN RSA PUBLIC KEY-----\n';
-  pem += `${der.match(/.{1,64}/g).join('\n')}`;
-  pem += '\n-----END RSA PUBLIC KEY-----\n';
-  return pem;
-}
+import jose from 'jose';
+import JwksError from './errors/JwksError';
 
 export function retrieveSigningKeys(keys) {
-  return keys
-    .filter((key) => {
-      if(key.kty !== 'RSA') {
-        return false;
-      }
-      if(key.hasOwnProperty('use') && key.use !== 'sig') {
-        return false;
-      }
-      return ((key.x5c && key.x5c.length) || (key.n && key.e));
-    })
-    .map(key => {
-      const jwk = {
-        kid: key.kid,
-        alg: key.alg,
-        nbf: key.nbf
-      };
-      const hasCertificateChain = key.x5c && key.x5c.length;
-      if (hasCertificateChain) {
-        jwk.publicKey = certToPEM(key.x5c[0]);
-        jwk.getPublicKey = () => jwk.publicKey;
-      } else {
-        jwk.rsaPublicKey = rsaPublicKeyToPEM(key.n, key.e);
-        jwk.getPublicKey = () => jwk.rsaPublicKey;
-      }
-      return jwk;
-    });
+  let keystore = [];
+  try {
+    keystore = jose.JWKS.asKeyStore({ keys }, { ignoreErrors: true });
+  } catch (err) {
+    return cb(new JwksError(err.message));
+  }
+  return keystore.all({ use: 'sig' }).map((key) => {
+    return {
+      kid: key.kid,
+      alg: key.alg,
+      get publicKey() { return key.toPEM(false); },
+      get rsaPublicKey() { return key.toPEM(false); },
+      getPublicKey() { return key.toPEM(false); }
+    };
+  });
 }
