+++ tests/mocks/jwks.js
@@ -10,9 +10,6 @@ export function jwksEndpoint(host, certs) {
         return {
           ...parsed,
           use: 'sig',
-          x5c: [
-            /-----BEGIN CERTIFICATE-----([^-]*)-----END CERTIFICATE-----/g.exec(cert.pub)[1].replace(/[\n|\r\n]/g, '')
-          ],
           kid: cert.kid
         };
       })
